refactor(FundraiseStatus): Change format of FundraiseStatus Enum
